add proposal adapter
