+++ packages/app/pages/beneficiary-proposals/takedowns/[id].tsx
@@ -2,7 +2,8 @@ import ProposalPage from 'components/Proposals/ProposalPage';
 import { ContractsContext } from 'context/Web3/contracts';
 import { Proposal } from 'interfaces/interfaces';
 import { useContext, useEffect, useState } from 'react';
-import { getProposal } from 'utils/getProposals';
+import { IpfsClient } from 'utils/IpfsClient';
+import { ProposalAdapter } from 'utils/ProposalAdapter';
 
 export default function SingleTakedownPage(): JSX.Element {
   const { contracts } = useContext(ContractsContext);
@@ -13,7 +14,9 @@ export default function SingleTakedownPage(): JSX.Element {
   }, []);
   useEffect(() => {
     if (contracts) {
-      getProposal(contracts, proposalId).then((res) => setProposal(res));
+      ProposalAdapter(contracts.beneficiaryGovernance, IpfsClient)
+        .getProposal(proposalId)
+        .then((res) => setProposal(res));
     }
   }, [contracts]);
   return <ProposalPage {...proposal} />;
