+++ packages/app/pages/beneficiary-proposals/takedowns/index.tsx
@@ -2,15 +2,17 @@ import { useContext, useEffect, useState } from 'react';
 import ProposalGrid from 'components/Proposals/ProposalGrid';
 import { ContractsContext } from 'context/Web3/contracts';
 import { Proposal } from 'interfaces/interfaces';
-import { getProposals } from 'utils/getProposals';
+import { IpfsClient } from 'utils/IpfsClient';
+import { ProposalAdapter } from 'utils/ProposalAdapter';
 
 export default function TakedownPage(): JSX.Element {
   const { contracts } = useContext(ContractsContext);
   const [takedownProposals, setTakedownProposals] = useState<Proposal[]>([]);
-
   useEffect(() => {
     if (contracts) {
-      getProposals(contracts, 1).then((res) => setTakedownProposals(res));
+      ProposalAdapter(contracts.beneficiaryGovernance, IpfsClient)
+        .getProposals(1)
+        .then((res) => setTakedownProposals(res));
     }
   }, [contracts]);
